requirements.txt
Pillow
matplotlib
configparser
opencv-python
numpy
